+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * `wallet_refresh_txs(wallet_id)` - refreshes all pending transactions of the given wallet
     * `wallet_reclaim_tx(wallet_id, tx_id)` - reclaims the funds from the given transaction
 * Add `id` to Transaction Response
+* Rename `CashedIn` to `Settled` (breaking DB Change)
 
 # 0.7.1